requirements.txt
# dev
beautifulsoup4==4.6.0
celery==4.1.0
Django==1.11.4
djangorestframework==3.6.4
django-allauth==0.33.0
django-celery-email==2.0.0
django-environ==0.4.4
django-extensions==1.9.0
django-rest-auth==0.9.1
psycopg2==2.7.3.1


# test
coverage==4.4.1
django-nose==1.4.5
factory-boy==2.9.2
gunicorn==19.7.1
ipdb==0.10.3
nose-timer==0.7.0
flake8==3.4.1
selenium==3.5.0
whitenoise==3.3.0
